insertion sort and reordered the algorithms folder
