+++ Algorithms/Searching/KMP algorithm.cpp
